backend/requirements.txt
fastapi
uvicorn[standard]
ypy-websocket
y-py
google-cloud-storage
google-cloud-datastore
ulid-py
